Refactor and improve motion profiling implementation
The motion profiling code was significantly refactored, providing better separation in functions and making the code cleaner. An option to set the polynomial's order in the CalcCamPolynomial method was added which provides more flexibility. Unnecessary lines and test functions were removed, and overall code quality was improved. New methods for evaluating and differentiating the CamProfile objects were introduced, further expanding the functionality.